% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_question.r
\name{annotation_question}
\alias{annotation_question}
\title{Create an annotation nquestion}
\usage{
annotation_question(
  name,
  question,
  codes = NULL,
  selection = c("buttons", "dropdown", "annotinder"),
  only_if = list(),
  not_if = list()
)
}
\arguments{
\item{name}{The name/label of the question}

\item{codes}{The codes that the coder can choose from. Can be a character vector, named character vector or data.frame.
An unnamed character vector creates codes with random colors.
A named character vector uses the names as codes and the values as colors, either as HEX or a name recognized by browsers (see \url{https://www.w3schools.com/colors/colors_names.asp}).
A data.frame must have a code column, and can use certain special columns (see details).}

\item{selection}{The method for selecting codes/answers. Can be "buttons", "dropdown" or "annotinder".
"buttons" shows all answers as buttons, "dropdown" gives a dropdown menu with a search bar.
"annotinder" lets users swipe for answers, and can only be used if the number of answers is 2 or 3.
The direction for each answer is "left" (first answer), "right" (second answer) and "up" (optional third answer).}

\item{only_if}{(TODO!!) Conditions that need to be met for this question to be asked. Given as a list, where the names are the names of other questions
in the codebook, and the values are character vectors with codes in this question. If the conditions are not met, this question will automatically
be answered with the code "IRRELEVANT". For example: list(relevant = "yes") means: Only ask this question if the question named "relevant" was answered with "yes". Can use multiple conditions:
list(relevant = "yes", mood = c("angry", "sad")) means that relevant needs to be answered "yes" and mood needs to be answered with either "angry" or "sad".}

\item{not_if}{(TODO!!) Like only_if, but for negative conditions. list(relevant = "no") would mean: don't ask this question if the question named "relevant" was
answered with the code "no".}

\item{instruction}{A short (think 1 or 2 sentences) question.}
}
\value{
A question object, to be used within the \code{\link{create_codebook}} function
}
\description{
Creates a question that can be passed as an argument to \code{\link{create_codebook}}.
}
\details{
Using a data.frame for the codes argument gives more flexibility. This data.frame should have a "code" column, and can in addition have a "color" and "parent" column
The color should be a color name, either as HEX or a name recognized by browsers (see \url{https://www.w3schools.com/colors/colors_names.asp})
The parent column is only relevant if you have many codes and use selection="dropdown". The dropdown menu will then show the codes with parent names,
and parent names are included in the search string. A parent can be the name of another code, and parents can have parents,
thus creating trees (just make sure not to create cycles). Use case would for example be an ontology with actor -> government -> president, and issue -> economy -> taxes.
}
