% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_annotator.r
\name{annotator_client}
\alias{annotator_client}
\title{Open the annotator client}
\usage{
annotator_client(in_browser = FALSE)
}
\arguments{
\item{in_browser}{Can be TRUE if you want to force opening in the default webbrowser}
}
\value{
Nothing
}
\description{
Opens the annotator client. If you're running RStudio, it will try to use the Viewer.
Otherwise, it will open in your default webbrowser.
}
\examples{
annotator_client()
}
