% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.r
\name{bind_codes}
\alias{bind_codes}
\title{Create a codes data.frame}
\usage{
bind_codes(...)
}
\arguments{
\item{...}{codes as created with \code{\link{code}}. Can also be a list of
codes}
}
\value{
a data.frame of codes
}
\description{
Create a codes data.frame
}
\examples{
# plain list of codes
codes <- c("here", "some", "codes")

# codes with extra arguments
codelist <- list(
  code("yes", color = "green"),
  code("no", color = "red")
)
bind_codes(codelist)

# codes with children
codelist <- c(
  code(
    "birds",
    code("chicken"),
    code("dove")
  ),
  code(
    "mammals",
    code("dog"),
    code("bear")
  )
)
}
