% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{c.codingjobUnit}
\alias{c.codingjobUnit}
\title{S3 print method for combining codingjobUnits}
\usage{
\method{c}{codingjobUnit}(...)
}
\arguments{
\item{...}{one or multiple codingjobUnit objects of codingjobUnits lists}
}
\description{
S3 print method for combining codingjobUnits
}
