% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.r
\name{code}
\alias{code}
\title{Create a code}
\usage{
code(code, ..., color = NULL)
}
\arguments{
\item{code}{The name of the code}

\item{...}{Optionally, nested codes for children}

\item{color}{Optionally, the color of the code}
}
\value{
A annotatorCode object, to be used inside the \code{\link{codes}} function
}
\description{
Create a code
}
\examples{
code('yes', color='green')
code('no', color='red')

# nested codes
code('actor',
    code('government',
        code('president'),
        code('vice-president')
    ),
    code('media'),
    code('society')
)
}
