% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_variable.r
\name{codebook_variable}
\alias{codebook_variable}
\title{Create a variable for annotate mode in \code{\link{create_codebook}}}
\usage{
codebook_variable(
  name,
  instruction,
  codes,
  selection = "buttons",
  onlyEdit = F,
  multiple = F
)
}
\arguments{
\item{name}{The name/label of the variable}

\item{instruction}{A brief (think 1 or 2 sentences) instruction to the coder.}

\item{codes}{The codes that the coder can choose from. Can be a character vector.
For more detailed settings (custom colors, nesting) use the \code{\link{codes}}}

\item{selection}{The method for selecting codes. Can be "buttons", "dropdown" or "dropdown_with_recent".
"buttons" shows all codes as a button, "dropdown" gives a dropdown menu with a search bar, and
"dropdown_with_recent" gives the dropdown with in addition buttons for recently used codes}

\item{onlyEdit}{If TRUE, coders can only edit existing annotations. !! Note that this requires \code{\link{create_codebook}} to have
have imported annotations for this variable.}

\item{multiple}{If TRUE, coder can select multiple codes for a selected piece of text before closing the popup. Note that they can always select
multiple codes by opening the popup multiple times. The setting exists for cases where multiple codes for a selection are common (e.g. topics in a paragraph).}
}
\value{
A variable object, to be used within the \code{\link{create_codebook}} function
}
\description{
Create a variable for annotate mode in \code{\link{create_codebook}}
}
\examples{
# simple variable with simple codes
codebook_variable("topic", "Assign topics to words",
                  codes = c("Economy","War","Health"))

# using codes and code functions for detailed codes
x = codebook_variable("sentiment", "Assign sentiment to words",
codes(
  code('negative', color='red'),
  code('neutral', color='grey'),
  code('positive', color='green')
  )
)

# get codes from a data.frame
codes_df = data.frame(code  = c("negative","neutral","positive"),
                      color = c("red",     "grey",   "green"))
codebook_variable("sentiment", "Assign sentiment to words", codes_df)
}
