% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.r
\name{codes}
\alias{codes}
\title{Create a codes data.frame}
\usage{
codes(...)
}
\arguments{
\item{...}{unnamed arguments, each one a code. Can either be a character like "yes",
or a more detailed code created with the \code{\link{code}} function.}
}
\value{

}
\description{
Create a codes data.frame
}
\examples{
# plain list of codes
codes('here','some','codes')

# codes with extra arguments
codes(
   code('yes', color='green'),
   code('no', color='red')
)

# codes with children
codes(
  code('birds',
     code('chicken'),
     code('dove')
   ),
  code('mammals',
     code('dog'),
     code('bear')
   )
)
}
