% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_backend.r
\name{conn_from_env}
\alias{conn_from_env}
\title{Get the current connection}
\usage{
conn_from_env()
}
\value{
an annotatorBackendConnection connection object, created with \link{backend_connect}
}
\description{
Get the current connection
}
