% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.r
\name{create_codebook}
\alias{create_codebook}
\title{Create a codebook for the CCS Annotator}
\usage{
create_codebook(mode = c("annotate", "questions"), ...)
}
\arguments{
\item{mode}{The type of annotation task. Currently supports "annotate" and "questions"}

\item{...}{Questions for "questions" mode or Variables for "annotate" mode. See \code{\link{codebook_variable}}}
}
\value{
A codebook object
}
\description{
Create a codebook for the CCS Annotator
}
\examples{
create_codebook('annotate',
codebook_variable("sentiment", "Assign sentiment to words", codes(
  code('negative', color='red'),
  code('neutral', color='grey'),
  code('positive', color='green')
)))
}
