% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_units.r
\name{create_info_unit}
\alias{create_info_unit}
\title{Create a special unit for informing coders}
\usage{
create_info_unit(markdown, button = "Continue")
}
\arguments{
\item{markdown}{Alternatively, provide the content as a markdown string. If
used, title+text are ignored. See \url{https://commonmark.org/help/} for
help.}

\item{button}{The text to show on the button}

\item{title}{A title}

\item{text}{The information text}
}
\value{
An informing unit
}
\description{
Create a special unit for informing coders
}
