% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_job.r
\name{create_job}
\alias{create_job}
\title{Create a CCS Annotator codingjob}
\usage{
create_job(title, units, codebook)
}
\arguments{
\item{title}{A character string, for the title of the codingjob}

\item{units}{A codingjobUnits object, as created with \code{\link{create_units}}}

\item{codebook}{A codebook object, as created with \code{\link{create_codebook}}}
}
\value{
A codingjob object
}
\description{
Create a CCS Annotator codingjob
}
