% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_units.r
\name{create_question_unit}
\alias{create_question_unit}
\title{Create a special unit for asking questions}
\usage{
create_question_unit(title, text, ..., text_window_size = NULL)
}
\arguments{
\item{title}{A title}

\item{text}{Some text to introduce the}

\item{...}{Questions to ask, created with the \code{\link{question}}}

\item{text_window_size}{The annotation interface has two parts: text (on top) and the answer form (at the bottom). By default, the text window size is about 70 (70%),
and users can change the size themselves (because the best size might differ depending on device). Here you can optionally set a fixed alternative.
This can for instance be useful if you have a question with big or many buttons, in which case you could set the text window size to something small.}
}
\value{

}
\description{
Sometimes you don't want to ask a coder about a unit, but rather about themselves, or something.
For instance, to include some basic survey questions at the start of a coding set.
}
