% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debrief.r
\name{debrief}
\alias{debrief}
\title{Debrief coder}
\usage{
debrief(message, link = NULL)
}
\arguments{
\item{message}{A message to show when a coder finishes a job.}

\item{link}{A url that will  be shown when the job is finished. If the url contains {user_id}, this will be replaced with the user ID (an email address or ID assigned via jobtoken)}
}
\description{
Debrief a coder when a job is finished.
}
