% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.r
\name{delete_token}
\alias{delete_token}
\title{Delete backend API token}
\usage{
delete_token()
}
\value{
Nothing, just deletes the thing
}
\description{
Just in case you really want to.
}
\examples{
delete_token()
}
