% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.r
\name{do_create_units}
\alias{do_create_units}
\title{#' Create gold units
#'
#'
#'
#' @param gold
#' @inheritParams create_units
#'
#' @return
#' @export
#'
#' @examples
#' d = data.frame(id = c(1,2),
#'                title = c('title 1', 'title 2'),
#'                text=c('text 1','text 2'),
#'                sentiment=c('positive','negative'))
#'
#' create_units(d, id='id', text=c('title','text'), gold='sentiment')
create_gold_units <- function(d, id, gold, text=NULL, meta=NULL, image=NULL, markdown=NA, variables=NULL) {
  do_create_units(d, id, text=text, meta=meta, image=image, markdown=markdown, variables=variables, gold=gold)
}}
\usage{
do_create_units(
  d,
  id,
  text = NULL,
  meta = NULL,
  image = NULL,
  markdown = NA,
  variables = NULL,
  gold = NULL
)
}
\description{
#' Create gold units
#'
#'
#'
#' @param gold
#' @inheritParams create_units
#'
#' @return
#' @export
#'
#' @examples
#' d = data.frame(id = c(1,2),
#'                title = c('title 1', 'title 2'),
#'                text=c('text 1','text 2'),
#'                sentiment=c('positive','negative'))
#'
#' create_units(d, id='id', text=c('title','text'), gold='sentiment')
create_gold_units <- function(d, id, gold, text=NULL, meta=NULL, image=NULL, markdown=NA, variables=NULL) {
  do_create_units(d, id, text=text, meta=meta, image=image, markdown=markdown, variables=variables, gold=gold)
}
}
