% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_management.r
\name{download_annotations}
\alias{download_annotations}
\title{Download annotation for a given job_id}
\usage{
download_annotations(job_id)
}
\arguments{
\item{job_id}{The codingjob id.}
}
\value{
A tibble of annotations in long format
}
\description{
Download annotation for a given job_id
}
\examples{
\dontrun{
download_annotations(7)
}
}
