% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme_annotations.r
\name{gimme_annotations}
\alias{gimme_annotations}
\title{Get annotation for a given job DB}
\usage{
gimme_annotations(db_file = NULL, only_done = FALSE)
}
\arguments{
\item{db_file}{If NULL (default) looks whether an an annotation server (see
\code{\link{start_annotator}}) ran (or is running) in the current session.
If so, it retrieves the annotations from this server}

\item{only_done}{If TRUE (default) only retrieve annotations with the "DONE"
status. If FALSE, also retrieve annotations with the "IN_PROGRESS" status}
}
\value{
A tibble with annotations. The id column matches with the units (as
  specified in the id argument in create_units)
}
\description{
Get annotation for a given job DB
}
\examples{
# get annotation for most recent server run (or still running) in current session
gimme_annotations()

\dontrun{
# from a job database, as returned by create_job_db (or by start_annotator)
job_db = "path/to/job.db"
gimme_annotations(job_db)
}
}
