% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{image_field}
\alias{image_field}
\title{Set image content}
\usage{
image_field(field, base64 = FALSE, caption = NULL, ...)
}
\arguments{
\item{field}{The content of the field. Can be given as a single filename, the name of a column in data (for create_units) that has filenames, or an expression.}

\item{base64}{If TRUE, store the image as a base64 in the codingjob json file}

\item{caption}{The image caption. Can be a single string, or a column in data (for create_units), or an expression.}

\item{...}{Style settings, passed to \code{\link{style}}}
}
\value{
Only meant to be used inside of \code{\link{create_units}} or \code{\link{create_unit}}.
}
\description{
Set image content
}
