% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_management.r
\name{list_jobs}
\alias{list_jobs}
\title{Get a list of codingjobs}
\usage{
list_jobs()
}
\value{
A data.frame with the codingjobs
}
\description{
Get the codingjobs for the current host and user, logged in with \code{\link{backend_connect}}
}
