% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mini_sotu_par}
\alias{mini_sotu_par}
\title{A few State of the Union addresses per paragraph}
\format{
data.frame
}
\usage{
data(mini_sotu_par)
}
\description{
A few State of the Union addresses per paragraph
}
\keyword{datasets}
