% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_backend.r
\name{print.annotatorBackendConnection}
\alias{print.annotatorBackendConnection}
\title{S3 print method for annotatorBackendConnection (API connection) objects}
\usage{
\method{print}{annotatorBackendConnection}(x, ...)
}
\arguments{
\item{x}{an annotatorBackendConnection connection object, created with \link{backend_connect}}

\item{...}{not used}
}
\description{
S3 print method for annotatorBackendConnection (API connection) objects
}
\examples{
\dontrun{
conn = backend_connect('http://localhost:8000')
conn
}
}
