% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.r
\name{print.codebook}
\alias{print.codebook}
\title{S3 print method for codebook objects}
\usage{
\method{print}{codebook}(x, ...)
}
\arguments{
\item{x}{an codebook object, created with \link{create_codebook}}

\item{...}{not used}
}
\description{
S3 print method for codebook objects
}
