% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_variable.r
\name{print.codebookVariable}
\alias{print.codebookVariable}
\title{S3 print method for codebookVariable objects}
\usage{
\method{print}{codebookVariable}(x, ...)
}
\arguments{
\item{x}{an codebookVariable object, created with \link{variable}}

\item{...}{not used}
}
\description{
S3 print method for codebookVariable objects
}
