% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_variable.r
\name{print.codeRelation}
\alias{print.codeRelation}
\title{S3 print method for codeRelation objects}
\usage{
\method{print}{codeRelation}(x, ...)
}
\arguments{
\item{x}{a codeRelation object, created with \link{relation}}

\item{...}{not used}
}
\description{
S3 print method for codeRelation objects
}
