% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.codingjobUnit}
\alias{print.codingjobUnit}
\title{S3 print method for codingjobUnit objects}
\usage{
\method{print}{codingjobUnit}(x, ...)
}
\arguments{
\item{x}{an codingjobUnit object}

\item{...}{not used}
}
\description{
S3 print method for codingjobUnit objects
}
