% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{print.createUnitsBundle}
\alias{print.createUnitsBundle}
\title{S3 print method for createUnitsBundle objects}
\usage{
\method{print}{createUnitsBundle}(x, ...)
}
\arguments{
\item{x}{an createUnitsBundle object, created with \link{create_units}}

\item{...}{not used}
}
\description{
S3 print method for createUnitsBundle objects
}
