% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.r
\name{print.textFields}
\alias{print.textFields}
\title{S3 print method for textFields objects}
\usage{
\method{print}{textFields}(x, ...)
}
\arguments{
\item{x}{an textFields object, created with \link{variable}}

\item{...}{not used}
}
\description{
S3 print method for textFields objects
}
