% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_grid}
\alias{set_grid}
\title{Customize the layout of the unit fields}
\usage{
set_grid(data, areas, rows = NULL, columns = NULL)
}
\arguments{
\item{areas}{A list with character vectors to specify the grid-template-areas (see \url{https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-areas}).
Each item in the list represents a row, and each value in the character vector the column in that row.
The values need to be the names of fields (i.e. the column names). If you want a position in the grid to be empty, use a dot ".".}

\item{rows}{A numeric vector of the same length as the number of rows. Each value indicates the relative space given to the row. So c(1,2) means that
the second row will be twice as high as the first row.}

\item{columns}{Same as rows, but for columns.}
}
\value{

}
\description{
By default, the fields are presented in a single column, in the given order (i.e. the order of set_text, set_markdown and set_image functions).
}
