% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_markdown}
\alias{set_markdown}
\title{Set markdown content}
\usage{
set_markdown(data, column, ...)
}
\arguments{
\item{data}{A createUnitsBundle object, as created with \code{\link{create_units}}}

\item{column}{The name of a column in data that contains markdown strings}

\item{...}{Style settings, passed to \code{\link{style}}}
}
\value{
A createUnitsBundle object
}
\description{
Set markdown content
}
