% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_question}
\alias{set_question}
\title{Create a unit specific codebook}
\usage{
set_question(name, question = NULL, codes = NULL, ...)
}
\arguments{
\item{name}{A character value indicating the name of the question. This will
also be the variable name in annotations. Coders won't see this name.}

\item{question}{The question text. Can either be a character value, or an
expression (e.g., to use a column in the data)}

\item{codes}{The codes that the coder can choose from. See
\code{\link{question}} for more details. Note that with set_question you
can refer to columns in the data to create dynamic questions.}

\item{...}{Other arguments passed to \code{\link{question}}}
}
\value{
Only meant to be used inside of \code{\link{create_units}} or
  \code{\link{create_unit}}.
}
\description{
Codebooks can be defined at different levels: codingjob > jobset > unit. The
most specific codebook will be used. This allows creating special units that
have their own codebook (e.g., for survey-like questions), or using codebooks
with dynamic codes.
}
