% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_text}
\alias{set_text}
\title{Set text content}
\usage{
set_text(
  data,
  column = NULL,
  before = NULL,
  after = NULL,
  label = NULL,
  ...,
  offset = 0
)
}
\arguments{
\item{data}{A createUnitsBundle object, as created with \code{\link{create_units}}}

\item{column}{The name of a "character" column in the data that contains coding unit text.
Note that this can also be empty if the current text field is only context (before or after the coding unit).
For example, the data.frame could be a keyword in context listing with the columns "pre", "keyword" and "post" (see for instance quanteda's kwic function).
These could then be set to the "before", "column" and "after" arguments, respectively.}

\item{before}{The text can have a context before and after the coding unit. This context will be shown to coders in grey,
and they cannot annotate it (in annotate mode). The 'before' argument can be the name of a "character" column. If specified,
the value of this column will be included in the current text_field as context. NOTE that if a text_field has a 'before' context, all text_fields before it
will automatically also be considered as context. (i.e. context can only occur before of after the coding unit, not within it)}

\item{after}{See 'before' argument}

\item{label}{A character value to label the text field. Coders will then see this label where this field starts.}

\item{...}{Style settings, passed to \code{\link{style}}}

\item{offset}{If the text is a part of a bigger text, you can include the offset for the character position where it starts. This is relevant if you want to import
or export span annotations}
}
\value{
A createUnitsBundle object
}
\description{
Set text content
}
