% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_server.r
\name{start_annotator}
\alias{start_annotator}
\title{Start annotation server (RStudio only)}
\usage{
start_annotator(codingjob, db_path = getwd(), overwrite = F, browse = T)
}
\arguments{
\item{codingjob}{A codingjob, created with \code{\link{create_job}}}

\item{db_path}{The path where the folder with coding job DBs is stored. Default is working directory.
If you're just playing around, tempdir() is pretty solid.}

\item{overwrite}{You're not allowed to create two jobs with the same title (which also becomes the DB filename).
Or well, you're allowed to, but you have to say overwrite is TRUE so you can't blame us if you
accidentally delete any hard-earned annotations.}

\item{browse}{If TRUE (default), automatically opens your default browser to start the annotation.}
}
\value{

}
\description{
Starts a server as a background job. While the server is running you can still
work in the current session and retrieve the annotations. The job is saved as a
SQLite DB in the ccsAnnotatorJobs folder (see db_path). So you can continue also retrieve
results or continue a job after closing (or breaking) the current R session.
}
