% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_variable.r
\name{summary.codebookVariable}
\alias{summary.codebookVariable}
\title{S3 summary method for codebookVariable objects}
\usage{
\method{summary}{codebookVariable}(object, ...)
}
\arguments{
\item{object}{an codebookVariable object, created with \link{annotation_variable}}

\item{...}{not used}
}
\description{
S3 summary method for codebookVariable objects
}
