% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.r
\name{text_field}
\alias{text_field}
\title{Create a detailed text_field for \code{\link{create_units}}}
\usage{
text_field(
  coding_unit = NULL,
  context_before = NULL,
  context_after = NULL,
  sep = " ",
  label = NULL,
  bold = F,
  italic = F,
  size = 1,
  justify = T,
  center = F,
  paragraphs = T,
  style = list(),
  offset = 0
)
}
\arguments{
\item{coding_unit}{The name of a "character" column in the "data" argument in \code{\link{create_units}} that contains coding unit text.
 Note that this can also be empty if the current text field is only context (before or after the coding unit).
 For example, the data.frame could be a keyword in context listing with the columns "pre", "keyword" and "post" (see for instance quanteda's kwic function).
These could then be set to the "context_before", "coding_unit" and "context_after" arguments, respectively.}

\item{context_before}{The text can have a context before and after the coding unit. This context will be shown to coders in grey,
and they cannot annotate it (in annotate mode). The context_before argument can be the name of a "character" column. If specified,
the value of this column will be included in the current text_field as context. NOTE that if a text_field has a context_before, all text_fields before it
will automatically also be considered as context. (i.e. context can only occur before of after the coding unit, not within it)}

\item{context_after}{See context_before.}

\item{sep}{A character string to separate the coding_unit from the context_before and context_after. Can also be a vector of length 2 for
different separators before and after.}

\item{label}{A character value to label the text field. Coders will then see this label where this field starts.}

\item{bold}{If TRUE, make the text field bold}

\item{italic}{If TRUE, make the text field bold}

\item{size}{A scaling value for the text size. default is 1. A value of 2 would be twice the size. A value of 0.5 half the size.}

\item{justify}{If TRUE (default) justify the text}

\item{center}{If TRUE, center the text}

\item{paragraphs}{If TRUE (default) show line breaks}

\item{style}{You can also directly specify inline CSS}

\item{offset}{If the text is a part of a bigger text, you can include the offset for the character position where it starts. This can
be relevant for connecting annotations at specific character positions between the full text and this text_field.}
}
\value{
Nothing. Should only be used inside of the \code{\link{create_units}} function
}
\description{
This function can be used inside the \code{\link{create_units}} function for more
customization options for text fields.
}
