% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.r
\name{text_fields}
\alias{text_fields}
\title{Helper function for text_field customization in \code{\link{create_units}}}
\usage{
text_fields(...)
}
\arguments{
\item{...}{unnamed arguments, each one a text_field. Can either be a character like "title" or "text",
or a more detailed text_field created with the \code{\link{text_field}} function.}
}
\value{
A textFields object, that can be passed to the "text" argument in \code{\link{create_units}}
}
\description{
Helper function for text_field customization in \code{\link{create_units}}
}
\examples{
d = data.frame(id=c(1,2), title = c('title 1', 'title 2'), text=c('text 1','text 2'),
pre=c('<','<'), post=c('>','>'), date=c('2010','2020'), topic=c('a','b'))

## simple units
create_units(d, id='id', text=c('title','text'), meta='date', variable='topic')

## using text_fields() for more options
create_units(d, 'id',
             text_fields(text_field('title'),
                         text_field('text', context_before='pre', context_after='post'))
)
}
