% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_management.r
\name{upload_job}
\alias{upload_job}
\title{Upload a CCS Annotator codingjob to a server}
\usage{
upload_job(codingjob)
}
\arguments{
\item{codingjob}{A codingjob created with \code{\link{create_job}}}
}
\value{
The id of the new codingjob on the server
}
\description{
Upload a CCS Annotator codingjob to a server
}
