% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_units.r
\name{create_info_unit}
\alias{create_info_unit}
\title{Create a special unit for informing coders}
\usage{
create_info_unit(title, text)
}
\arguments{
\item{title}{A title}

\item{text}{The information text}
}
\value{

}
\description{
Create a special unit for informing coders
}
