% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{create_unit}
\alias{create_unit}
\title{Create a single unit}
\usage{
create_unit(id, ..., type = "code")
}
\arguments{
\item{id}{A unique id}

\item{...}{}

\item{type}{The unit type. Can be 'code', 'test', 'train' or 'survey'}
}
\description{
Works like \code{\link{create_units}}, but for a single unit.
The values can then directly be provided in the set_ functions.
}
