% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.r
\name{create_units}
\alias{create_units}
\title{Create units}
\usage{
create_units(d, ..., text = NULL, meta = NULL, variables = NULL)
}
\arguments{
\item{d}{A data.frame}

\item{...}{Optionally, pass text_fields as (unnamed) arguments. If used, the text argument is ignored}

\item{text}{A character vector with column names to use as text fields, or the output of \code{\link{text_fields}}
for more customization options. text_field an also be passed as unnamed arguments (see ...)}

\item{meta}{A character vector with column names to show as a table of meta data above the units.}

\item{variables}{A character vector with column names to use as unit variables}
}
\value{
A list of units, to be used inside \code{\link{create_codebook}}
}
\description{
Create units
}
\examples{
d = data.frame(title = c('title 1', 'title 2'),
               text=c('text 1','text 2'),
               pre=c('<','<'), post=c('>','>'),
               date=c('2010','2020'),
               topic=c('a','b'))

## simple units
create_units(d, text=c('title','text'), meta='date', variables='topic')

## using text_fields() for more options
create_units(d, text_fields(
   text_field('title'),
   text_field('text', context_before='pre', context_after='post')
))
}
