% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{create_units}
\alias{create_units}
\title{Create units for the cssAnnotator}
\usage{
create_units(data, id, variables = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{id}{A column in the data.frame with unique values.}

\item{variables}{A vector of column names. These column names can then be referenced from the codebook.
For example, if there is a column "topic", you could ask the question: "is this sentence about {topic}".
The {topic} part will then be replaced in the question with the value for this unit in the "topic" column.}
}
\description{
Create the basis for a units object. The content of the units can then be
designed in a pipe of \code{\link{set_text}}, \code{\link{set_meta}}, \code{\link{set_image}} and \code{\link{set_markdown}} functions
}
