% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debrief.r
\name{debrief}
\alias{debrief}
\title{Debrief coder}
\usage{
debrief(message, link = NULL, link_text = NULL, qr = FALSE)
}
\arguments{
\item{message}{A message in markdown to show when a coder finishes a job. The text is centered}

\item{link}{A url that will  be shown when the job is finished. If the url contains "{user_id}", this will be replaced with the Username}

\item{link_text}{the text shown on the button with the link}

\item{qr}{If TRUE, show QR code for creating a new job coder at the end, so coders can share the job with others}
}
\description{
Debrief a coder when a job is finished.
}
