% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.r
\name{image_field}
\alias{image_field}
\title{Create a detailed text_field for \code{\link{create_units}}}
\usage{
image_field(image_file, caption = NULL)
}
\arguments{
\item{caption}{}
}
\value{

}
\description{
Create a detailed text_field for \code{\link{create_units}}
}
