% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.r
\name{print.codebookQuestion}
\alias{print.codebookQuestion}
\title{S3 print method for codebookQuestion objects}
\usage{
\method{print}{codebookQuestion}(x, ...)
}
\arguments{
\item{x}{an codebookQuestion object, created with \link{question}}

\item{...}{not used}
}
\description{
S3 print method for codebookQuestion objects
}
