% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.codingjobUnits}
\alias{print.codingjobUnits}
\title{S3 print method for codingjobUnits objects}
\usage{
\method{print}{codingjobUnits}(x, ...)
}
\arguments{
\item{x}{an codingjobUnits object, created with \link{create_units}}

\item{...}{not used}
}
\description{
S3 print method for codingjobUnits objects
}
