% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_markdown}
\alias{set_markdown}
\title{Set markdown content}
\usage{
set_markdown(data, column, split = NULL, ...)
}
\arguments{
\item{data}{A createUnitsBundle object, as created with \code{\link{create_units}}}

\item{column}{The name of a column in data that contains markdown strings}

\item{split}{A string for splitting the markdown string. The field will then be split over multiple markdown fields. See the per_field argument in \code{\link{question}}
for a cool way to use this to repeat a question for multiple parts of a text.}

\item{...}{Style settings, passed to \code{\link{style}}}
}
\value{
A createUnitsBundle object
}
\description{
Set markdown content
}
