% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_markdown}
\alias{set_markdown}
\title{Set markdown content}
\usage{
set_markdown(name, value, ...)
}
\arguments{
\item{...}{Style settings, passed to \code{\link{style}}}

\item{field}{The content of the field. Can be given as a single string, the name of a column in data (for create_units), or any expression.}
}
\value{
Only meant to be used inside of \code{\link{create_units}} or \code{\link{create_unit}}.
}
\description{
Set markdown content
}
