% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_meta}
\alias{set_meta}
\title{Set meta-data content}
\usage{
set_meta(data, column, label = NULL, bold = TRUE, ...)
}
\arguments{
\item{data}{A createUnitsBundle object, as created with \code{\link{create_units}}}

\item{column}{The column with the meta data.}

\item{label}{A character value to label the meta field.}

\item{bold}{Meta data by default uses 'bold' style setting.}

\item{...}{Style settings, passed to \code{\link{style}}}
}
\value{
A createUnitsBundle object
}
\description{
Set meta-data content
}
