% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_train}
\alias{set_train}
\title{Set training units}
\usage{
set_train(data, select)
}
\arguments{
\item{data}{A createUnitsBundle object, as created with \code{\link{create_units}}}

\item{select}{A logical expression to select which units are training units. Perhaps the best approach is to
already have a column in the data that indicates the type of units (code, train or test). The select
expression would then be something like: type == 'train'.}

\item{...}{\code{\link{condition}} functions, to indicate}
}
\description{
Set training units
}
