% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_units.r
\name{set_train}
\alias{set_train}
\title{Set test units}
\usage{
set_train(
  data,
  column,
  message = NULL,
  submessage = NULL,
  variable = column,
  damage = 0,
  operator = "=="
)
}
\arguments{
\item{data}{A createUnitsBundle object, as created with \code{\link{create_units}}}

\item{column}{The name of a column in data that has the correct annotation values. The name of the column needs to correspond to a variable/question
in the codebook, or alternatively use the 'variable' argument. If you need to set training answers for multiple columns (e.g.,
for units with multiple questions) you can use the set_train function multiple times.}

\item{message}{A markdown string that will be displayed when a coder gives an incorrect answer. If not given, the message will be:
\code{"### You gave an incorrect answer.\n\nThis is a **training** unit. \nPlease have another look and select a different answer"}.}

\item{submessage}{Optionally, The name of a column in data with a specific markdown string to display when this condition is not met.
This message will be displayed beneath the on_wrong message.}

\item{variable}{The name of the variable as used in the codebook. If not specified, the name of the column will be used.}

\item{damage}{The amount of damage a coder should receive.}

\item{operator}{How should the annotation value be compared to the column value? Default is "==" (equals). Alternatives are "!=" (not equals),
"<=", "<", ">=" or ">".}
}
\value{

}
\description{
Setup training units, and provide the correct answers/annotations for these units.
If the answer/annotation is wrong, coders will see a message and need to retry.
Note that which units are train units needs to be indicated with the 'type' argument in \code{\link{create_units}}
}
