% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.r
\name{summary.codebookQuestion}
\alias{summary.codebookQuestion}
\title{S3 summary method for codebookQuestion objects}
\usage{
\method{summary}{codebookQuestion}(x, ...)
}
\arguments{
\item{x}{an codebookQuestion object, created with \link{variable}}

\item{...}{not used}
}
\description{
S3 summary method for codebookQuestion objects
}
